swap examples
