+++ library/sdk-examples/example-widget-dynamic-refs-recursive.ts
@@ -1,38 +1,67 @@
-app({
-    ui: (form) => {
-        // 🔶 if you want to use a form dynamically,
-        // for now, you need to form.shared first aroudn it
-        const listOfStuff = form.shared(
-            'test1',
-            form.list({
-                label: 'Sampler',
-                defaultLength: 3,
-                min: 1,
-                element: () => form.string(),
-            }),
-        )
+import type { Widget, Widget_selectOne } from 'src'
+import type { Widget_choices } from 'src/controls/widgets/choices/WidgetChoices'
+import type { Widget_group } from 'src/controls/widgets/group/WidgetGroup'
+import type { Widget_image } from 'src/controls/widgets/image/WidgetImage'
+import type { Widget_list } from 'src/controls/widgets/list/WidgetList'
 
-        return {
-            listOfStuff,
-            listOfRefs: form.list({
-                label: 'Sampler',
-                defaultLength: 3,
-                min: 1,
-                element: () =>
-                    form.selectOne({
-                        label: 'dynamic-test',
-                        choices: () => {
-                            const CHOICES = listOfStuff.shared.items.map((item, ix) => ({
-                                id: item.serial.id,
-                                label: item.value,
-                            }))
-                            console.log(`[🔴] ready`, CHOICES)
-                            return CHOICES
-                        },
-                    }),
-            }),
-        }
-    },
+app({
+    ui: (form) => ({
+        samplerUI: form.list({
+            label: 'Sampler',
+            // showID: true,
+            startCollapsed: true,
+            defaultLength: 1,
+            min: 1,
+            element: () =>
+                form.choice({
+                    items: {
+                        sampler_output_abc_asdf: () =>
+                            form.selectOne({
+                                // showID: true,
+                                // if choices is a function, the form root is injected as first parameter
+                                choices: (formRoot: Widget_group<any>) => {
+                                    // 🔶 null when the form is not yet fully initialized
+                                    if (formRoot.fields.samplerUI == null) return []
+
+                                    // 🔶 self-referencing => typescript can't infer the type here
+                                    const steps = formRoot.fields.samplerUI as Widget_list<
+                                        Widget_choices<{
+                                            sampler_output_abc_asdf: () => Widget_selectOne<any>
+                                            empty_latent: () => Widget_group<any>
+                                            pick_image: () => Widget_image
+                                        }>
+                                    >
+
+                                    // return a list of items
+                                    return steps.items.map((choiceWidget: Widget_choices<any>, ix: number) => {
+                                        if (choiceWidget == null) console.log(`[🔴] err 1: choiceWidget is null`)
+                                        const _selectOne = choiceWidget.firstActiveBranchWidget as Maybe<Widget_selectOne<any>>
+                                        if (_selectOne == null) console.log(`[🔴] err 2: firstActiveBranchWidget is null`, _selectOne) // prettier-ignore
+                                        const _actualChoice = _selectOne?.value
+                                        return {
+                                            id: _selectOne?.id ?? 'error',
+                                            disabled: _actualChoice == null,
+                                            name: _selectOne?.type ?? '❌ ERROR',
+                                            label: `${ix + 1}th (${choiceWidget.firstActiveBranchName ?? '❓'})`,
+                                        }
+                                    })
+                                },
+                            }),
+                        empty_latent: () =>
+                            form.group({
+                                layout: 'H',
+                                // topLevel: true,
+                                items: () => ({
+                                    width: form.int({ default: 512, max: 1512, step: 32, hideSlider: true }),
+                                    height: form.int({ default: 512, max: 1512, step: 32, hideSlider: true }),
+                                    batch: form.int({ default: 1, min: 1, max: 32, hideSlider: true }),
+                                }),
+                            }),
+                        pick_image: () => form.image({}),
+                    },
+                }),
+        }),
+    }),
     run(run, ui) {
         console.log(`[🟢] done`)
     },
